.azure-pipelines: specify which macos projects to build
The change to explicitly pass `--runtime=osx-x64` to dotnet uncovered a
couple existing issues that were previously obfuscated. The second issue,
which this change addresses, is that we cannot build an entire solution
with a specific runtime; instead, we must explicitly specify which
project(s) we want to build with that runtime.