macOS release: sign payload
Add step two of the macOS release process to download the unsigned payload,
use ESRP to sign it, and upload the signed payload as an artifact.

Additionally, genericize the existing run_esrp_signing.py script to allow
for signing of macOS payload/pkg and Windows exes in addition to Debian
packages.